+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v0.3.0] 2018-08-24
+### Added
+- `Projectionist` to play, boot and retire projectors
+- `ProjectorStateLedger` that tracks projector state information
+- `Projection`, `Projector` and `EventHandlingStrategy` interfaces
+- `EventHandlingListener` projectionist
+
 ## [v0.2.0] 2018-08-21
 ### Added
 - `Event` interface
@@ -17,10 +24,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [v0.1.0] 2018-08-20
 ### Added
-- Basic documentation and static project files.
+- Basic documentation and static project files
 - Initial composer definitions.
 - PHPSpec bootstrap
 
-[Unreleased]: https://github.com/slickframework/cqrs-tools/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/slickframework/cqrs-tools/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/slickframework/cqrs-tools/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/slickframework/cqrs-tools/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/slickframework/cqrs-tools/compare/85b339f...v0.1.0